# [RFC] optimize get_values for dictionaries