This method is about twice as fast in my testing on dictionaries, and
just slightly faster overall.

```
root@9f5e81940873:/# bash old.sh; bash old.sh; bash old.sh
10 loops, best of 3: 1.15 sec per loop
10 loops, best of 3: 1.13 sec per loop
10 loops, best of 3: 1.12 sec per loop
root@9f5e81940873:/# !!:gs/old/new/
bash new.sh; bash new.sh; bash new.sh
10 loops, best of 3: 564 msec per loop
10 loops, best of 3: 571 msec per loop
10 loops, best of 3: 555 msec per loop
```

This is our code dumping 5000 objects comparing the current get_values
vs the one I am proposing.

And this is the marshmallow bench tests.

```
dev $ python performance/benchmark.py; python performance/benchmark.py; python performance/benchmark.py;
Benchmark Result: 839.86 usec/dump
Benchmark Result: 851.05 usec/dump
Benchmark Result: 839.13 usec/dump
dev $ git checkout gtmanfred-dev dev
Switched to branch 'gtmanfred-dev'
Your branch is ahead of 'origin/dev' by 1 commit.
  (use "git push" to publish your local commits)
gtmanfred-dev $ python performance/benchmark.py; python performance/benchmark.py; python performance/benchmark.py;
Benchmark Result: 852.94 usec/dump
Benchmark Result: 811.79 usec/dump
Benchmark Result: 815.60 usec/dump
gtmanfred-dev $ python performance/benchmark.py; python performance/benchmark.py; python performance/benchmark.py;
Benchmark Result: 824.43 usec/dump
Benchmark Result: 819.98 usec/dump
Benchmark Result: 856.79 usec/dump
```