@gtmanfred 2019-01-04
yeah, turns out my benchmarks were fast because they were always returning the defaults, sorry about the noise.